Testy-

+++ src/main/java/com/dawkastresu/medicalclinic/service/DoctorService.java
@@ -3,18 +3,19 @@ package com.dawkastresu.medicalclinic.service;
 import com.dawkastresu.medicalclinic.command.RegisterDoctorCommand;
 import com.dawkastresu.medicalclinic.dto.DoctorDto;
 import com.dawkastresu.medicalclinic.exception.DoctorNotFoundException;
-import com.dawkastresu.medicalclinic.model.Doctor;
-import com.dawkastresu.medicalclinic.model.Institution;
-import com.dawkastresu.medicalclinic.model.Password;
+import com.dawkastresu.medicalclinic.model.*;
 import com.dawkastresu.medicalclinic.repository.DoctorRepository;
 import com.dawkastresu.medicalclinic.repository.InstitutionRepository;
 import com.dawkastresu.medicalclinic.utils.DoctorMapper;
 import com.dawkastresu.medicalclinic.utils.DoctorValidator;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.time.LocalDate;
 
 @Service
 @RequiredArgsConstructor
@@ -24,21 +25,36 @@ public class DoctorService {
     private final InstitutionRepository institutionRepository;
     private final DoctorMapper mapper;
 
-    public List<DoctorDto> getAll() {
-        return repository.findAll().stream()
-                .map(mapper::toDto)
-                .toList();
+    public Page<DoctorDto> getAll(Pageable pageable) {
+        Page<Doctor> page = repository.findAll(pageable);
+        return page.map(mapper::toDto);
     }
 
+    public Page<DoctorDto> getDoctorsBySpecialization(Specialization specialization, Pageable pageable) {
+        return repository
+                .findDoctorBySpecialization(specialization, pageable)
+                .map(mapper::toDto);
+    }
+
+    @Transactional
     public DoctorDto addNew(RegisterDoctorCommand command) {
         Doctor doctor = Doctor.create(command);
+
         if (DoctorValidator.validateDoctor(repository, doctor.getFirstName())) {
-            doctor = repository.save(doctor); // Przypisujemy zwrócony obiekt do zmiennej doctor
+            doctor = repository.save(doctor);
+        } else {
+            throw new IllegalArgumentException("Invalid doctor data");
         }
+
         return mapper.toDto(doctor);
     }
 
+
+    @Transactional
     public void remove(Long id) {
+        if (!repository.existsById(id)) {
+            throw new DoctorNotFoundException("No doctor found with this id", HttpStatus.NOT_FOUND);
+        }
         repository.deleteById(id);
     }
 
@@ -48,6 +64,7 @@ public class DoctorService {
         return mapper.toDto(doctor);
     }
 
+    @Transactional
     public DoctorDto editById(Long id, RegisterDoctorCommand command) {
         Doctor doctor = repository.findById(id)
                 .orElseThrow(() -> new DoctorNotFoundException("Doctor not found", HttpStatus.NOT_FOUND));
@@ -56,7 +73,6 @@ public class DoctorService {
         DoctorValidator.newValueNotNullValidate(newDoctor);
         DoctorValidator.validateDoctorEdit(newDoctor);
 
-        doctor.setInstitutions(institutionRepository.findAllById(command.getInstitutionIds()));
         doctor.update(newDoctor);
         repository.save(doctor);
 
@@ -71,7 +87,10 @@ public class DoctorService {
     public void addDoctorToInstitution(Long idDoctor, Long idInstitution) {
         Institution institution = institutionRepository.findById(idInstitution).orElseThrow(() -> new IllegalArgumentException());
         Doctor doctor = repository.findById(idDoctor).orElseThrow(() -> new IllegalArgumentException());
-        institution.setDoctors(List.of(doctor));
+        doctor.getInstitutions().add(institution);
+        institution.getDoctors().add(doctor);
+        repository.save(doctor);
+        institutionRepository.save(institution);
     }
 
 }
